update(praktikum - 1/Pembagian.java): change `double` to `int` at `banyak_apel` variable

+++ pertemuan-1/Pembagian.java
@@ -1,6 +1,6 @@
 class Pembagian {
     public static void main(String[] args) {
-        int banyak_apel  = 5;
+        double banyak_apel  = 5;
         int jumlah_anak  = 2;
         // int perolehan   = banyak_apel / jumlah_anak;
         double perolehan = banyak_apel / jumlah_anak;
